Amend tsconfig to correctly exclude files, and separate cypress and project imports
